2-add_node.c

+++ 0x12-singly_linked_lists/2-add_node.c
@@ -12,7 +12,7 @@
 
 list_t *add_node(list_t **head, const char *str)
 {
-	size_t len = _strlen(str);
+	int len = _strlen(str);
 	list_t *node = malloc(sizeof (list_t));
 
 	if (!node)
@@ -33,9 +33,9 @@ list_t *add_node(list_t **head, const char *str)
  * Return: length of string
  */
 
-size_t _strlen(char *str)
+int _strlen(char *str)
 {
-	size_t i = 0;
+	int i = 0;
 
 	while (*str)
 		i++;
